requirements.txt
anyio==4.12.1
APScheduler==3.6.3
asyncpg==0.31.0
blinker==1.9.0
cachetools==4.2.2
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
Flask==3.1.3
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-telegram-bot==22.6
pytz==2025.2
redis==7.2.0
setuptools==82.0.0
six==1.17.0
SQLAlchemy==2.0.46
tornado==6.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
Werkzeug==3.1.6
alembic==1.13.2
